+++ src/main/java/io/kontur/eventapi/dao/FeedDao.java
@@ -42,7 +42,7 @@ public class FeedDao {
                 feedData.getName(), feedData.getProperName(), feedData.getDescription(),
                 feedData.getStartedAt(), feedData.getEndedAt(), feedData.getUpdatedAt(),
                 feedData.getObservations(), episodesJson, feedData.getEnriched(), feedData.getUrls(),
-                feedData.getLocation());
+                feedData.getLocation(), feedData.getLatestSeverity(), feedData.getSeverities());
 
         mapper.markOutdatedEventsVersions(feedData.getEventId(), feedData.getFeedId(), feedData.getVersion());
         feedEventStatusDao.markAsActual(feedData.getFeedId(), feedData.getEventId(), true);