+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 #### Added
 
+- Added `severities` and `latest_severity` columns into `feed_data`
+
 #### Changed
 
 #### Removed